Update EnrichedEventFacet to use FacetQuery; update tests
