+++ src/cbc_sdk/endpoint_standard/__init__.py
@@ -3,4 +3,4 @@
 from __future__ import absolute_import
 
 from cbc_sdk.endpoint_standard.base import Device, Event, Policy, EnrichedEvent, Query, EnrichedEventQuery, \
-                                           EnrichedEventFacet, EnrichedEventFacetQuery 
+                                           EnrichedEventFacet
