+++ src/tests/unit/endpoint_standard/test_endpoint_standard_enriched_events_facet.py
@@ -2,8 +2,9 @@
 
 import pytest
 import logging
-from cbc_sdk.endpoint_standard import EnrichedEventFacet, EnrichedEventFacetQuery
+from cbc_sdk.endpoint_standard import EnrichedEventFacet
 from cbc_sdk.rest_api import CBCloudAPI
+from cbc_sdk.base import FacetQuery
 from cbc_sdk.errors import ObjectNotFoundError, ApiError
 from tests.unit.fixtures.CBCSDKMock import CBCSDKMock
 from tests.unit.fixtures.endpoint_standard.mock_enriched_events_facet import (POST_ENRICHED_EVENTS_FACET_SEARCH_JOB_RESP,
@@ -38,9 +39,9 @@ def test_enriched_event_facet_select_where(cbcsdk_mock):
 
     api = cbcsdk_mock.api
     events = api.select(EnrichedEventFacet).where(process_name="chrome.exe").add_facet_field("process_name")
-    for event in events:
-        assert event.terms is not None
-        assert event.ranges is not None
+    event = events.results
+    assert event.terms is not None
+    assert event.ranges is not None
 
 def test_enriched_event_facet_select_async(cbcsdk_mock):
     """Testing EnrichedEvent Querying with select()"""
@@ -60,9 +61,11 @@ def test_enriched_event_facet_select_compound(cbcsdk_mock):
 
     api = cbcsdk_mock.api
     events = api.select(EnrichedEventFacet).where(process_name="chrome.exe").or_(process_name="firefox.exe").add_facet_field("process_name")
-    for event in events:
-        assert event.terms is not None
-        assert event.ranges is not None
+    # for event in events:
+    #     assert event.terms is not None
+    #     assert event.ranges is not None
+    event = events.results
+    assert event.terms_.fields == ["process_name"]
 
 def test_enriched_event_facet_query_implementation(cbcsdk_mock):
     """Testing EnrichedEvent querying with where()."""
@@ -72,8 +75,9 @@ def test_enriched_event_facet_query_implementation(cbcsdk_mock):
     api = cbcsdk_mock.api
     field = 'process_name'
     events = api.select(EnrichedEventFacet).where(process_name="tesst").add_facet_field("process_name")
-    assert isinstance(events, EnrichedEventFacetQuery)
-    assert events[0].terms[0]["field"] == field
+    assert isinstance(events, FacetQuery)
+    event = events.results
+    assert event.terms[0]["field"] == field
 
 def test_enriched_event_facet_timeout(cbcsdk_mock):
     """Testing EnrichedEventQuery.timeout()."""
@@ -150,11 +154,11 @@ def test_enriched_event_facet_limit(cbcsdk_mock):
     api = cbcsdk_mock.api
     events = api.select(EnrichedEventFacet).where(process_pid=1000).limit(123).add_facet_field("process_name")
     assert events._limit == 123
-    
+
 def test_enriched_event_facet_time_range(cbcsdk_mock):
     """Testing EnrichedEvent results sort."""
     api = cbcsdk_mock.api
-    events = api.select(EnrichedEventFacet).where(process_pid=1000).set_time_range(start="2020-10-10T20:34:07Z", 
+    events = api.select(EnrichedEventFacet).where(process_pid=1000).set_time_range(start="2020-10-10T20:34:07Z",
                                                                               end="2020-10-20T20:34:07Z",
                                                                               window="-1d").add_facet_field("process_name")
     assert events._time_range["start"] == "2020-10-10T20:34:07Z"
@@ -186,8 +190,9 @@ def test_enriched_events_search(cbcsdk_mock):
 
     api = cbcsdk_mock.api
     events = api.select(EnrichedEventFacet).where(process_pid=1000).add_facet_field("process_name")
-    events._search()
-    assert events[0].terms is not None
+    future = events.execute_async()
+    results = future.result()
+    assert results[0].terms is not None
 
 def test_enriched_events_search_async(cbcsdk_mock):
     """ Test _search method of enrichedeventquery class """
@@ -196,6 +201,6 @@ def test_enriched_events_search_async(cbcsdk_mock):
 
     api = cbcsdk_mock.api
     events = api.select(EnrichedEventFacet).where(process_pid=1000).add_facet_field("process_name")
-    events._search_async()
-    assert events[0].terms is not None
-    
+    future = events.execute_async()
+    results = future.result()
+    assert results[0].terms is not None
